change download trigger position, if no dev env installed, ask the user if install
