+++ app/src/main/java/com/termux/app/TermuxActivity.java
@@ -6,6 +6,7 @@ import android.annotation.TargetApi;
 import android.app.Activity;
 import android.app.AlertDialog;
 import android.app.Dialog;
+import android.app.ProgressDialog;
 import android.content.ActivityNotFoundException;
 import android.content.BroadcastReceiver;
 import android.content.ClipData;
@@ -24,8 +25,10 @@ import android.graphics.Typeface;
 import android.media.AudioAttributes;
 import android.media.SoundPool;
 import android.net.Uri;
+import android.os.AsyncTask;
 import android.os.Build;
 import android.os.Bundle;
+import android.os.Environment;
 import android.os.Handler;
 import android.os.IBinder;
 import android.os.Vibrator;
@@ -71,13 +74,18 @@ import com.termux.terminal.TerminalSession.SessionChangedCallback;
 import com.termux.terminal.TextStyle;
 import com.termux.view.TerminalView;
 
+import java.io.BufferedInputStream;
+import java.io.DataOutputStream;
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
+import java.io.OutputStream;
 import java.net.NetworkInterface;
+import java.net.URL;
+import java.net.URLConnection;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.LinkedHashSet;
@@ -150,6 +158,167 @@ public final class TermuxActivity extends Activity implements ServiceConnection 
         new AudioAttributes.Builder().setContentType(AudioAttributes.CONTENT_TYPE_SONIFICATION)
             .setUsage(AudioAttributes.USAGE_ASSISTANCE_SONIFICATION).build()).build();
     int mBellSoundId;
+
+    public static final String SHARPAI_URL = "https://github.com/SharpAI/DeepCamera/releases/download/v1.0/sharpai-bin.tgz";
+
+    /**
+     * Async Task to download file from URL
+     */
+    private static class DownloadFile extends AsyncTask<String, String, String> {
+
+        private ProgressDialog progressDialog;
+        private String fileName;
+        private String folder;
+        private boolean isDownloaded;
+        private Activity activity = null;
+
+        public DownloadFile(final Activity activity) {
+            super();
+            this.activity = activity;
+        }
+
+        /**
+         * Before starting background thread
+         * Show Progress Bar Dialog
+         */
+        @Override
+        protected void onPreExecute() {
+            super.onPreExecute();
+            this.progressDialog = new ProgressDialog(activity);
+            this.progressDialog.setProgressStyle(ProgressDialog.STYLE_HORIZONTAL);
+            this.progressDialog.setCancelable(false);
+            this.progressDialog.show();
+        }
+
+        /**
+         * Downloading file in background thread
+         */
+        @Override
+        protected String doInBackground(String... f_url) {
+            int count;
+            try {
+                URL url = new URL(f_url[0]);
+                URLConnection connection = url.openConnection();
+                connection.connect();
+                // getting file length
+                int lengthOfFile = connection.getContentLength();
+
+                // input stream to read file - with 8k buffer
+                InputStream input = new BufferedInputStream(url.openStream(), 8192);
+
+                //Extract file name from URL
+                fileName = f_url[0].substring(f_url[0].lastIndexOf('/') + 1, f_url[0].length());
+
+                //External directory path to save file
+                folder = Environment.getExternalStorageDirectory() + File.separator + "sharpaidownload/";
+
+                //Create androiddeft folder if it does not exist
+                File directory = new File(folder);
+
+                if (!directory.exists()) {
+                    directory.mkdirs();
+                }
+
+                // Output stream to write file
+                OutputStream output = new FileOutputStream(folder + fileName);
+
+                byte data[] = new byte[1024];
+
+                long total = 0;
+
+                while ((count = input.read(data)) != -1) {
+                    total += count;
+                    // publishing the progress....
+                    // After this onProgressUpdate will be called
+                    publishProgress("" + (int) ((total * 100) / lengthOfFile));
+                    Log.d("sharpai", "Progress: " + (int) ((total * 100) / lengthOfFile));
+
+                    // writing data to file
+                    output.write(data, 0, count);
+                }
+
+                // flushing output
+                output.flush();
+
+                // closing streams
+                output.close();
+                input.close();
+                return "" + folder + fileName;
+
+            } catch (Exception e) {
+                Log.e("Error: ", e.getMessage());
+            }
+
+            return "Something went wrong";
+        }
+
+        /**
+         * Updating progress bar
+         */
+        protected void onProgressUpdate(String... progress) {
+            // setting progress percentage
+            progressDialog.setProgress(Integer.parseInt(progress[0]));
+        }
+
+
+        @Override
+        protected void onPostExecute(String downloadedFile) {
+            // dismiss the dialog after the file was downloaded
+            this.progressDialog.dismiss();
+
+            // Display File path after downloading
+            Toast.makeText(activity,
+                "downloaded to " + downloadedFile, Toast.LENGTH_LONG).show();
+            extractDownloadedFile(downloadedFile);
+        }
+
+        private void extractDownloadedFile(String downloadedFile) {
+            final ProgressDialog progress = ProgressDialog.show(activity, null, activity.getString(R.string.sharpai_extract_body), true, false);
+            new Thread() {
+                @Override
+                public void run() {
+                    try {
+                        try {
+                            String tarCmd = "busybox tar -xmf " + downloadedFile + " -C " + TermuxService.FILES_PATH + "/\n";
+                            Process untar = Runtime.getRuntime().exec(tarCmd);
+                            DataOutputStream outputStream = new DataOutputStream(untar.getOutputStream());
+
+                            outputStream.writeBytes("exit\n");
+                            outputStream.flush();
+                            untar.waitFor();
+                        }catch(IOException e){
+                            TermuxInstaller.deleteFile(downloadedFile);
+                            throw new Exception(e);
+                        }catch(InterruptedException e){
+                            TermuxInstaller.deleteFile(downloadedFile);
+                            throw new Exception(e);
+                        } finally {
+                            TermuxInstaller.deleteFile(downloadedFile);
+                            Log.i(">>>> lambda<<<< ", "解压成功！！！！");
+                        };
+                    } catch (final Exception e) {
+                        Log.e(EmulatorDebug.LOG_TAG, "sharpai error", e);
+                        activity.runOnUiThread(() -> {
+                            try {
+                                Toast.makeText(activity,
+                                    "extract file " + downloadedFile + "failed!", Toast.LENGTH_LONG).show();
+                            } catch (WindowManager.BadTokenException e1) {
+                                // Activity already dismissed - ignore.
+                            }
+                        });
+                    } finally {
+                        activity.runOnUiThread(() -> {
+                            try {
+                                progress.dismiss();
+                            } catch (RuntimeException e) {
+                                // Activity already dismissed - ignore.
+                            }
+                        });
+                    };
+                }
+            }.start();
+        }
+    }
 
     public class SharpAIRunnable implements Runnable {
         public final String CWD = HOME_PATH+"/DeepCamera";
@@ -393,6 +562,18 @@ public final class TermuxActivity extends Activity implements ServiceConnection 
             askIfRunDeepCameraService();
             return;
         }
+        else { //no dev env, ask if to download
+            new AlertDialog.Builder(this).setTitle(R.string.sharpai_install_title).setMessage(R.string.sharpai_install_msg)
+                .setNegativeButton(R.string.sharpai_dialog_abort, (dialog, which) -> {
+                    dialog.dismiss();
+                }).setPositiveButton(R.string.sharpai_dialog_ok, (dialog, which) -> {
+                dialog.dismiss();
+                installSharpAISystem();
+            }).show();
+        }
+    }
+    private void installSharpAISystem() {
+        new DownloadFile(this).execute(SHARPAI_URL);
     }
     private Boolean checkIfHasDeepCameraDevFile(){
         File file = new File(HOME_PATH+"/DeepCamera/start_service.sh");
